+++ .gitignore
@@ -1,3 +1,5 @@
 bower_components
+node_modules
+dist
 .idea
 *.iml
